### Enhance Query - Change amount type to Integer - Add query to fetchIncome & fetchOutcome

+++ app/src/main/java/com/example/moneytrack/CreateScreenActivity.kt
@@ -108,7 +108,7 @@ class CreateScreenActivity : AppCompatActivity() {
                     title = title,
                     description = description,
                     type = type,
-                    amount = amount,
+                    amount = amount.toInt(),
                     date = date,
                 )
             )
